Implemented customization of what value should be used for conversion.
When converting XP to Emeralds, the amount may now specify the amount
of XP, or the number of Emeralds desired. When converting from Emeralds
to XP, the amount may be the number of Emeralds, or the target
experience level that is desired. Both defined by respective options
in config file.